+++ kdl/org.integratedmodelling.kdl/src-gen/org/integratedmodelling/kdl/kdl/impl/UnitElementImpl.java
@@ -327,7 +327,7 @@ public class UnitElementImpl extends MinimalEObjectImpl.Container implements Uni
   {
     if (eIsProxy()) return super.toString();
 
-    StringBuffer result = new StringBuffer(super.toString());
+    StringBuilder result = new StringBuilder(super.toString());
     result.append(" (id: ");
     result.append(id);
     result.append(')');