+++ kim/org.integratedmodelling.kim/src-gen/org/integratedmodelling/kim/kim/impl/ClassifierRHSImpl.java
@@ -1163,7 +1163,7 @@ public class ClassifierRHSImpl extends MinimalEObjectImpl.Container implements C
   {
     if (eIsProxy()) return super.toString();
 
-    StringBuffer result = new StringBuffer(super.toString());
+    StringBuilder result = new StringBuilder(super.toString());
     result.append(" (boolean: ");
     result.append(boolean_);
     result.append(", leftLimit: ");