+++ kim/org.integratedmodelling.kim/src-gen/org/integratedmodelling/kim/kim/impl/DocSelectorImpl.java
@@ -385,7 +385,7 @@ public class DocSelectorImpl extends MinimalEObjectImpl.Container implements Doc
   {
     if (eIsProxy()) return super.toString();
 
-    StringBuffer result = new StringBuffer(super.toString());
+    StringBuilder result = new StringBuilder(super.toString());
     result.append(" (id: ");
     result.append(id);
     result.append(", definition: ");